requirements.txt
discord.py
openai
python-dotenv
requests
beautifulsoup4
lxml
pytest
playwright
sqlalchemy
firecrawl
pydantic
fire
